made delete-product endpoint safe

+++ pages/api/product/delete-product/[id].ts
@@ -1,15 +1,17 @@
 import prisma from "../../../../lib/prisma";
+import { getSession } from "next-auth/react";
 
 export default async function handle(req, res) {
     const postId = req.query.id;
-    if (req.method === "DELETE") {
+    const session = await getSession({ req })
+    if (req.method === "DELETE" && session) {
         const post = await prisma.post.delete({
             where: { id: postId },
         });
         res.json(post)
     } else {
         throw new Error(
-            `The HTTP ${req.method} is not supported at this route`
+            req.method !== "DELETE" ? `The HTTP ${req.method} is not supported at this route` : "You are not authorized"
         )
     }
 }
